requirements.txt
torch
transformers>=4.41.0
datasets>=2.19.0
evaluate>=0.4.2
accelerate>=0.30.0
pyyaml>=6.0.1
scikit-learn>=1.4.0
pytest>=7.0.0

